Meld MapViewController into AddPostController
