+++ MVP/MVP/Controllers/View Controllers/AddPost/AddPostViewController.swift
@@ -10,6 +10,8 @@ import UIKit.UIImage
 import Firebase
 import AVFoundation
 import Photos
+import MapKit
+import CoreLocation
 
 class AddPostViewController: UIViewController {
     
@@ -19,13 +21,15 @@ class AddPostViewController: UIViewController {
     @IBOutlet weak var currentUserNameLabel: UILabel!
     @IBOutlet weak var postImageView: UIImageView!
     @IBOutlet weak var currentUserProfileImageView: UIImageView!
+    @IBOutlet weak var mapView: MKMapView!
     
     //MARK: - Properties
     let db = Firestore.firestore()
-    
     let imagePicker = UIImagePickerController()
-    //landing pad for image grabbed in PhotoSelector, use this for saving to the cloud
     var selectedImage: UIImage?
+    let locationManager = CLLocationManager()
+    var currentUser = CurrentUserController.shared.currentUser
+    let fifteenMiles: Double = 24140.16
     
     //MARK: - Lifecycle
     override func viewDidLoad() {
@@ -33,10 +37,17 @@ class AddPostViewController: UIViewController {
         imagePicker.delegate = self
         titleTextField.delegate = self
         setUpViews()
+        let tap = UITapGestureRecognizer(target: self.view, action: #selector(UIView.endEditing))
+        view.addGestureRecognizer(tap)
+        currentUser?.location = locationManager.location
+        mapView.register(MKMarkerAnnotationView.self, forAnnotationViewWithReuseIdentifier: MKMapViewDefaultAnnotationViewReuseIdentifier)
+        
+        mapView.delegate = self
+        checkLocationServices()
     }
     
     func setUpViews()  {
-        currentUserNameLabel.text = CurrentUserController.shared.currentUser?.firstName
+        currentUserNameLabel.text = currentUser?.firstName
         if selectedImage != nil {
             postImageView?.image = selectedImage
             print("\npostImageView assigned")
@@ -54,9 +65,9 @@ class AddPostViewController: UIViewController {
         
         let postCreatedTimestamp = timestamp()
         
-        let currentUser = Auth.auth().currentUser
+        let authUser = Auth.auth().currentUser
         
-        guard let currentUserFirstName = CurrentUserController.shared.currentUser?.firstName, let postUserUID = currentUser?.uid else {return}
+        guard let currentUserFirstName = currentUser?.firstName, let postUserUID = authUser?.uid else {return}
         
         if selectedImage == nil {
             
@@ -68,6 +79,8 @@ class AddPostViewController: UIViewController {
 
             grabImageURLAndSavePost(postTitle: postTitle, postDescription: postDescription, postUserUID: postUserUID, postUserFirstName: currentUserFirstName, postCreatedTimestamp: postCreatedTimestamp)
         }
+        addRandomLocation()
+        print("Randomized location")
     }
     
     @IBAction func addPhotoButtonTapped(_ sender: UIButton) {
@@ -90,14 +103,18 @@ class AddPostViewController: UIViewController {
         present(alert, animated: true, completion: nil)
     }//end of addPhotoButtonTapped func
     
+    @IBAction func clearButtonTapped(_ sender: Any) {
+        clearAllFields()
+    }
+    
     //MARK: - Helpers
     func grabImageURLAndSavePost(postTitle: String, postDescription: String, postUserUID: String, postUserFirstName: String, postCreatedTimestamp: String) {
         
         guard let imageData = selectedImage!.jpegData(compressionQuality: 0.6) else {return}
         let storageReference = Storage.storage().reference()
-        let currentUser = Auth.auth().currentUser
-        guard let postUserUID = currentUser?.uid else {return}
-        let postImageRef = storageReference.child("users").child(currentUser!.uid).child("\(currentUser!.uid)-postImage.jpg")
+        let authUser = Auth.auth().currentUser
+        guard let postUserUID = authUser?.uid else {return}
+        let postImageRef = storageReference.child("users").child(authUser!.uid).child("\(authUser!.uid)-postImage.jpg")
         let uploadMetaData = StorageMetadata()
         uploadMetaData.contentType = "image/jpeg"
         
@@ -135,20 +152,20 @@ class AddPostViewController: UIViewController {
     //MARK: - TODO: move to PostController file
     func savePost(postTitle: String, postDescription: String, postUserUID: String, postUserFirstName: String, postCreatedTimestamp: String, imageURL: String = "none") {
         
-        let currentUser = Auth.auth().currentUser
+        let authUser = Auth.auth().currentUser
         
         let dateFormatter = DateFormatter()
         dateFormatter.dateFormat = "M/d/Y"
         let formattedDate = dateFormatter.string(from: Date())
         
-        guard let currentUserFirstName = CurrentUserController.shared.currentUser?.firstName, let currentUserLatitude = CurrentUserController.shared.currentUser?.latitude, let currentUserLongitude = CurrentUserController.shared.currentUser?.longitude else {return}
+        guard let currentUserFirstName = currentUser?.firstName, let currentUserLatitude = currentUser?.latitude, let currentUserLongitude = currentUser?.longitude else {return}
         
         var ref: DocumentReference? = nil
         
         let data: [String: Any] = [
             "postTitle": "\(postTitle)",
             "postDescription": "\(postDescription)",
-            "postUserUID": "\(currentUser!.uid)",
+            "postUserUID": "\(authUser!.uid)",
             "postUserFirstName": "\(currentUserFirstName)",
             "postUserLatitude": currentUserLatitude,
             "postUserLongitude": currentUserLongitude,
@@ -178,11 +195,6 @@ class AddPostViewController: UIViewController {
         self.titleTextField.text = ""
         self.descriptionTextView.text = ""
     }
-    
-    @IBAction func clearButtonTapped(_ sender: Any) {
-        clearAllFields()
-    }
-    
     
     func presentSavedAlert() {
         let title = "Post saved"
@@ -299,7 +311,7 @@ extension AddPostViewController: UIImagePickerControllerDelegate {
         }
     }//end of checkPhotoLibraryAuthorization func
     
-    func presentDeniedAlert() {
+    func presentPhotoDeniedAlert() {
         let alert = UIAlertController(title: "Access Denied", message: "Check your permission or restriction settings and try again.", preferredStyle: .alert)
         
         let dismissButton = UIAlertAction(title: "Okay", style: .default, handler: nil)
@@ -309,12 +321,9 @@ extension AddPostViewController: UIImagePickerControllerDelegate {
     }//end of presentDeniedAlert func
 }//end of AddPostTVC photoSelector extension
 
-
 extension AddPostViewController: UINavigationControllerDelegate {
     
 }
-
-
 
 //this dismisses the keyboard when enter is pressed on an iphone
 extension AddPostViewController: UITextFieldDelegate {
@@ -323,6 +332,104 @@ extension AddPostViewController: UITextFieldDelegate {
     }
 }
 
+extension AddPostViewController: MKMapViewDelegate, CLLocationManagerDelegate {
+    func checkLocationServices() {
+        if CLLocationManager.locationServicesEnabled() {
+            setupLocationManager()
+            checkLocationAuthorization()
+        } else {
+            locationServicesDisabledAlert()
+        }
+    }
+    
+    func setupLocationManager() {
+        locationManager.delegate = self
+        locationManager.desiredAccuracy = kCLLocationAccuracyBest
+    }
+    
+    func checkLocationAuthorization() {
+        switch CLLocationManager.authorizationStatus() {
+        case .authorizedWhenInUse:
+            mapView.showsUserLocation = true
+            centerViewOnUserLocation()
+        case .denied:
+            presentDeniedAlert()
+            break
+        case .notDetermined:
+            locationManager.requestWhenInUseAuthorization()
+            break
+        case .restricted:
+            presentRestrictedAlert()
+            break
+        case .authorizedAlways:
+            centerViewOnUserLocation()
+        @unknown default:
+            print("Default case for CLLocationManager.authorizationStatus()")
+        }
+    }//end of checkLocationAuthorization func
+
+    func centerViewOnUserLocation() {
+        if let location = locationManager.location?.coordinate {
+            let region = MKCoordinateRegion.init(center: location, latitudinalMeters: fifteenMiles, longitudinalMeters: fifteenMiles)
+            mapView.setRegion(region, animated: true)
+        }
+    }//end of centerViewOnUserLocation func
+    
+    func addRandomLocation() {
+        guard let currentUser = currentUser else {return}
+        let radius = 1200
+        let randomCenter = CurrentUserController.shared.generateRandomLocation(for: currentUser).coordinate
+        let circle = MKCircle(center: randomCenter, radius: CLLocationDistance(radius))
+        mapView.addOverlay(circle)
+    }
+
+     func mapView(_ mapView: MKMapView, rendererFor overlay: MKOverlay) -> MKOverlayRenderer {
+        if overlay.isKind(of: MKCircle.self){
+            let circleRenderer = MKCircleRenderer(overlay: overlay)
+            circleRenderer.fillColor = UIColor.blue.withAlphaComponent(0.2)
+            circleRenderer.strokeColor = UIColor.blue
+            circleRenderer.lineWidth = 1
+            return circleRenderer
+        }
+        return MKOverlayRenderer(overlay: overlay)
+    }
+    
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+           checkLocationAuthorization()
+           
+           let userLocation: CLLocation = locations[0] as CLLocation
+
+           CurrentUserController.shared.currentUser?.location = userLocation
+
+       }
+    
+     func locationServicesDisabledAlert() {
+        let alert = UIAlertController(title: "Location Services Disabled", message: "Check your location service settings and try again.", preferredStyle: .alert)
+
+         let dismissButton = UIAlertAction(title: "Okay", style: .default, handler: nil)
+
+         alert.addAction(dismissButton)
+        self.present(alert, animated: true)
+    }//end of locationServicesDisabledAlert func
+    
+    func presentDeniedAlert() {
+        let alert = UIAlertController(title: "Access Denied", message: "Check your permission settings and try again.", preferredStyle: .alert)
+        
+        let dismissButton = UIAlertAction(title: "Okay", style: .default, handler: nil)
+        
+        alert.addAction(dismissButton)
+        self.present(alert, animated: true)
+    }//end of presentDenied Alert func
+    
+    func presentRestrictedAlert() {
+        let alert = UIAlertController(title: "Access Restricted", message: "Permissions (i.e. Parental Controls) have been disabled for this user.", preferredStyle: .alert)
+        
+        let dismissButton = UIAlertAction(title: "Okay", style: .default, handler: nil)
+        
+        alert.addAction(dismissButton)
+        self.present(alert, animated: true)
+    }//end of presentRestrictedAlert func
+}
 
 //trying to get the same thing for the textview but haven't figured it our yet
 //extension AddPostViewController: UITextViewDelegate {
